requirements.txt
# requirements.txt generated by rsconnect-python on 2022-11-30 20:53:29.794060
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
contextvars==2.4
contourpy==1.0.6
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
fonttools==4.38.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
htmltools==0.1.2
idna==3.4
immutables==0.19
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
linkify-it-py==2.0.0
locationtagger==0.0.1
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
murmurhash==1.0.9
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pathy==0.10.0
Pillow==9.3.0
preshed==3.0.8
progressbar==2.5
pycountry==22.3.5
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
seaborn==0.12.1
semver==2.13.0
sgmllib3k==1.0.0
shiny==0.2.9
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
starlette==0.22.0
thinc==8.1.5
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
ultimate-sitemap-parser==0.5
urllib3==1.26.13
uvicorn==0.20.0
wasabi==0.10.1
websockets==10.4
